Changed from X to + layout, increased max throttle
This configuration is now for my 35cm wheelbase quadcopter using my old
Park 300 motors. The new lightweight frame is suited for both flightmodes,
and I chose to fly this one in + configuration.
To get the maximum power out of the ESCs, I increased the maximum throttle
value to 1990, very close to the configured ESC limit of 2000.